# openapi/blerp.v1.yaml
openapi: 3.1.0
info:
  title: Blerp Identity Service API
  version: 1.0.0-draft
  contact:
    name: Blerp Support
    url: https://github.com/e6qu/blerp
    email: adi11235@gmail.com
  summary: REST + JSON interface for Blerp Identity Service
  description: >
    Clean-room OpenAPI contract describing the public REST API for user, session,
    organization, and webhook management. This specification is the authoritative
    contract for backend/frontend work and must exist locally before any detailed
    implementation plan is executed.
...
servers:
  - url: https://api.blerp.dev
    description: Production
  - url: https://staging-api.blerp.dev
    description: Staging
  - url: http://localhost:4000
    description: Local development (docker-compose)
security:
  - SecretKey: []
...
components:
  securitySchemes:
    SecretKey:
      type: http
      scheme: bearer
      bearerFormat: secret-key
      description: Server-to-server secret API key (sk_live_* / sk_test_*).
    PublishableKey:
      type: http
      scheme: bearer
      bearerFormat: publishable-key
      description: Client-side publishable API key (pk_live_* / pk_test_*).
    SessionToken:
      type: http
      scheme: bearer
      bearerFormat: session-token
      description: Session token returned after sign-in/up (`sess_*`).
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          minimum: 0
        next_cursor:
          type: string
        previous_cursor:
          type: string
...
    Organization:
      type: object
      required: [id, project_id, name]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        public_metadata:
          type: object
          additionalProperties: true
        private_metadata:
          type: object
          additionalProperties: true
        domains:
          type: array
          items:
            type: object
            properties:
              domain:
                type: string
              status:
                type: string
                enum: [pending, verified, rejected]
              verified_at:
                type: string
                format: date-time
        settings:
          type: object
          properties:
            allow_invites:
              type: boolean
            default_role:
              type: string
            mfa_policy:
              type: string
              enum: [optional, required, disabled]
        created_at:
          type: string
          format: date-time
...
paths:
...
  /v1/organizations:
    get:
      tags: [Organizations]
      summary: List organizations
      description: Returns a paginated list of organizations in the current project.
      operationId: listOrganizations
      security:
        - SecretKey: []
      parameters:
        - in: query
          name: project_id
          schema:
            type: string
        - in: query
          name: domain
          schema:
            type: string
        - in: query
          name: cursor
          schema:
            type: string
      responses:
        "200":
          description: Organizations list
          content:
            application/json:
              schema:
                type: object
                required: [data, total_count]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Organization"
                  # BUG-48/58: Clerk-canonical pagination field. SDK
                  # clients should read total_count; meta is a legacy
                  # alias kept for one release.
                  total_count:
                    type: integer
                    minimum: 0
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
    post:
      tags: [Organizations]
      summary: Create organization
      description: Creates a new organization along with optional role defaults.
      operationId: createOrganization
      security:
        - SecretKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              # BUG-212 (codex r62): project_id is OPTIONAL. The API
              # derives it from the auth context when omitted —
              # session user's first owned project, or the M2M token's
              # bound project. Explicit project_id still wins.
              required: [name]
              properties:
                project_id:
                  type: string
                  description: |
                    Project the new organization belongs to. Optional
                    — when omitted the API derives the project from
                    the caller's auth context (session user's first
                    owned project; M2M token's bound project).
                name:
                  type: string
                slug:
                  type: string
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /v1/organizations/{organization_id}:
    get:
      tags: [Organizations]
      summary: Retrieve organization
      description: Fetches a single organization including metadata and member counts.
      operationId: getOrganization
      security:
        - SecretKey: []
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    patch:
      tags: [Organizations]
      summary: Update organization settings
      description: Updates organization metadata, membership requirements, or billing settings.
      operationId: updateOrganization
      security:
        - SecretKey: []
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                metadata_public:
                  type: object
                  additionalProperties: true
                metadata_private:
                  type: object
                  additionalProperties: true
                settings:
                  type: object
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    delete:
      tags: [Organizations]
      summary: Delete an organization
      description: |
        Permanently deletes the organization. Requires `org:write` on the
        target organization (a membership row with that permission must
        exist for the caller). The RBAC layer evaluates first, so a
        request for an organization the caller cannot access — whether
        the row exists or not — receives 403, never 404.
      operationId: deleteOrganization
      security:
        - SecretKey: []
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Organization deleted
        "403":
          description: Caller lacks `org:write` on the target organization (or the organization does not exist; the two cases are not distinguished, by design, to avoid leaking existence to unauthorized callers)
...